+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Remove aim dependencies such as keras, pytorch and etc  (gorarakelyan)
+
 ## 1.1.0 Jan 12 2020
 - Update code diff tracking to be optional (gorarakelyan)
 - Add default False value to aim init function (gorarakelyan)